requirements.txt
Django==3.2.3
djangorestframework>=3.12.4, <3.13.0
psycopg2>=2.7.5, <2.8.0
pyhunter>=1.7.0, <1.8.0
clearbit>=0.1.7, <0.2.0
django-rest-swagger